requirement.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.1

# LangGraph + LangChain
langgraph==0.2.28
langchain==0.3.0
langchain-core==0.3.0
langchain-groq==0.2.0
langchain-google-genai==2.0.0

# Database
# psycopg2-binary==2.9.9  # Comment out for now due to pg_config requirement
sqlalchemy==2.0.35

# Redis + Celery
redis==5.0.8
celery[redis]==5.4.0
flower==2.0.1

# Utilities
pydantic==2.9.0
pydantic-settings==2.5.2

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2